+++ src/main/javascript/index.js
@@ -1,5 +1,5 @@
 
-export connect from './Core/connect';
+export createApp from './Core/createApp';
 export App from './Core/App';
 export * as AppEvents from './Core/AppEvents';
 export * as UIEvents from './Core/UIEvents';